Skip static template on vital starter
